Move serve_with_tls to tls.rs
